leaves wind movement working

+++ ex5_PEPSE/pepse/world/Block.java
@@ -7,7 +7,7 @@ import danogl.gui.rendering.Renderable;
 import danogl.util.Vector2;
 
 public class Block extends GameObject {
-    public static final int SIZE = 20;
+    public static final float SIZE = 20;
     public Block(Vector2 topLeftCorner, Renderable renderable) {
         super(topLeftCorner, Vector2.ONES.mult(SIZE), renderable);
         physics().preventIntersectionsFromDirection(Vector2.ZERO);
